Split the code into graph,  admin, guess, settings and the rest
