+++ src/components/index.jsx
@@ -1,4 +1,3 @@
-export { default as AdminRoute } from './AdminRoute';
 export { default as BingomasterRoute } from './BingomasterRoute';
 export { default as ProtectedRoute } from './ProtectedRoute';
 export { default as ScrollToTop } from './ScrollToTop';
